Django/1019 to many relationship/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-appconf==1.0.5
django-bootstrap5==22.1
django-cleanup==6.0.0
django-imagekit==4.1.0
pilkit==2.0
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
